#34 fix profile section item
