+++ JobCareer/Presentation/Screen/Profile/Root/ProfileDataSource.swift
@@ -13,8 +13,8 @@ extension ProfileUI {
                 return nil
             }
 
-            switch (section, item) {
-                case let (.main, .main(item)):
+            switch section {
+                case .main:
                     guard
                         let cell = collectionView.dequeueReusableCell(
                             withReuseIdentifier: MainCell.resourceName,
@@ -24,10 +24,13 @@ extension ProfileUI {
                         return UICollectionViewCell()
                     }
 
-                    cell.configure(item: item)
+                    if case let .main(item) = item {
+                        cell.configure(item: item)
+                    }
+
                     return cell
 
-                case let (.career, .career(item)):
+                case .career:
                     guard
                         let cell = collectionView.dequeueReusableCell(
                             withReuseIdentifier: CareerCell.resourceName,
@@ -37,10 +40,13 @@ extension ProfileUI {
                         return UICollectionViewCell()
                     }
 
-                    cell.configure(item: item)
+                    if case let .career(item) = item {
+                        cell.configure(item: item)
+                    }
+
                     return cell
 
-                case (.logout, .logout):
+                case .logout:
                     guard
                         let cell = collectionView.dequeueReusableCell(
                             withReuseIdentifier: LogoutCell.resourceName,
@@ -52,9 +58,6 @@ extension ProfileUI {
 
                     self.delegate.didLogoutButtonTapped(cell.buttonTapPublisher)
                     return cell
-
-                default:
-                    return nil
             }
         }
     }
